+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # lighter-flagger Change Log
 
+## v1.2.0 - 2017-01-08
+* **minor** Added `_settingReady` to prevent recursive ready emits.
+
 ## v1.1.0 - 2016-10-06
-* `minor` Added wait, unwait, waitFor and ready.
+* **minor** Added `wait`, `unwait`, `waitFor`, `ready` and `then`.
 
 ## v1.0.0 - 2015-10-20
-* `major` Community release.
+* **major** Community release.